backend/requirements.txt
Django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2==2.9.1
pyjwt==2.1.0
python-dateutil==2.8.2
channels==3.0.4
schedule==1.1.0